Feat: ANT-4878/4879 - RES > Import trajectory: modify checksum calculation

+++ src/main/java/com/rte_france/antares/datamanager_back/repository/model/ResTechnologyDistributionEntity.java
@@ -32,7 +32,7 @@ public class ResTechnologyDistributionEntity {
     private String pecdTechnology;
 
     @Column(name = "capacity_by_year")
-    private Integer capacityByYear;
+    private Double capacityByYear;
 
     @ManyToOne
     @JoinColumn(name = "trajectory_id")
